added utility function to build the image screenshot of a project
